requirements.txt
# 网站推荐系统依赖包
PyQt6>=6.4.0
psycopg2-binary>=2.9.0